examples/composio-demos/slack-poster/requirements.txt
composio-core==0.7.12
